[ refactor ] Recover Intro_WhyInScope
