+++ src/full/Agda/Interaction/JSONTop.hs
@@ -90,7 +90,10 @@ instance ToJSON DisplayInfo where
     [ "kind"        .= String "SearchAbout"
     , "payload"     .= Null
     ]
-  toJSON (Info_WhyInScope doc) = object [ "kind" .= String "WhyInScope", "payload" .= render doc ]
+  toJSON (Info_WhyInScope _ _ _ _ _) = object
+    [ "kind"        .= String "WhyInScope"
+    , "payload"     .= Null
+    ]
   toJSON (Info_NormalForm doc) = object [ "kind" .= String "NormalForm", "payload" .= render doc ]
   toJSON (Info_GoalType doc) = object [ "kind" .= String "GoalType", "payload" .= render doc ]
   toJSON (Info_CurrentGoal doc) = object [ "kind" .= String "CurrentGoal", "payload" .= render doc ]
